however, changing back to u32 works

+++ lib.rs
@@ -1,8 +1,8 @@
 use wasm_bindgen::prelude::*;
 
-/// add two u64 numbers
+/// add two u32 numbers
 #[wasm_bindgen]
-pub fn add(a: u64, b: u64) -> u64 {
+pub fn add(a: u32, b: u32) -> u32 {
   return a + b;
 }
